Make `pip.list` use and return normalized package names
Fix the handling of Python modules with non-standard characters in
either the package name or the name of the installed module.
Passing the module name through the normalize() function
recommended by the Python packaging guidelines allows us to
correctly match them, so that e.g. the consumers do not falsely
report that a package was not installed correctly or needs to
be installed.

NB: this commit CHANGES the dictionary returned by `pip.list`;
the package names will be normalized, so any consumers that look for
specific names MUST also normalize the latter.